Create and Delete are done

+++ DVD Library/JS/home.js
@@ -16,28 +16,26 @@
 //     }
 // });
 
-$("#search-button").click( function() {
+$("#search-button").click(function () {
 
     $("#dvd-results-table").empty();
     $("#dvd-results-table").append("<tr><th>Title</th><th>Release Date</th><th>Director</th><th>Rating</th><th></th></tr>").text();
-    if($("dvds").val() == "search category" || $("#search").val() == "" || $("#search").val() == "Search Term")
-    {
+    if ($("dvds").val() == "search category" || $("#search").val() == "" || $("#search").val() == "Search Term") {
         $("#search-validation").css("visibility", "visible");
     }
-    else
-    {
+    else {
         $("#search-validation").css("visibility", "hidden");
-        if($("#dvds").val() == "title") {
+        if ($("#dvds").val() == "title") {
             var search = $("#search").val();
             $.ajax({
                 type: "GET",
                 url: "http://localhost:8080/dvds/title/" + search,
                 contentType: "application/json;charset=UTF-8",
                 success: function (dvdArray) {
-                    $.each(dvdArray, function(index, dvd) {
+                    $.each(dvdArray, function (index, dvd) {
                         $("#dvd-results-table").append("<tr id="
                             + dvd.dvdId
-                            + "><td><button type='button' onclick=displayDVD("+dvd.dvdId+")>"
+                            + "><td><button type='button' onclick=displayDVD(" + dvd.dvdId + ")>"
                             + dvd.title
                             + "</button></td><td>"
                             + dvd.realeaseYear
@@ -45,26 +43,26 @@ $("#search-button").click( function() {
                             + dvd.director
                             + "</td><td>"
                             + dvd.rating
-                            + "</td><td><button type='button' onclick=displayEdit("+dvd.dvdId+")>Edit</button><button type='button' onclick=deleteDVD("+dvd.dvdId+")>Delete</button></td></tr>").text();
+                            + "</td><td><button type='button' onclick=displayEdit(" + dvd.dvdId + ")>Edit</button><button type='button' onclick=deleteDVD(" + dvd.dvdId + ")>Delete</button></td></tr>").text();
                     })
                 },
-                error: function(xhr, status, error) {
+                error: function (xhr, status, error) {
                     alert(xhr.responseText);
                 }
             });
         }
 
-        if($("#dvds").val() == "release year") {
+        if ($("#dvds").val() == "release year") {
             var search = $("#search").val();
             $.ajax({
                 type: "GET",
                 url: "http://localhost:8080/dvds/year/" + search,
                 contentType: "application/json;charset=UTF-8",
                 success: function (dvdArray) {
-                    $.each(dvdArray, function(index, dvd) {
+                    $.each(dvdArray, function (index, dvd) {
                         $("#dvd-results-table").append("<tr id="
                             + dvd.dvdId
-                            + "><td><button type='button' onclick=displayDVD("+dvd.dvdId+")>"
+                            + "><td><button type='button' onclick=displayDVD(" + dvd.dvdId + ")>"
                             + dvd.title
                             + "</button></td><td>"
                             + dvd.realeaseYear
@@ -72,26 +70,26 @@ $("#search-button").click( function() {
                             + dvd.director
                             + "</td><td>"
                             + dvd.rating
-                            + "</td><td><button type='button' onclick=displayEdit("+dvd.dvdId+")>Edit</button><button type='button' onclick=deleteDVD("+dvd.dvdId+")>Delete</button></td></tr>").text();
+                            + "</td><td><button type='button' onclick=displayEdit(" + dvd.dvdId + ")>Edit</button><button type='button' onclick=deleteDVD(" + dvd.dvdId + ")>Delete</button></td></tr>").text();
                     })
                 },
-                error: function(xhr, status, error) {
+                error: function (xhr, status, error) {
                     alert(xhr.responseText);
                 }
             });
         }
 
-        if($("#dvds").val() == "director name") {
+        if ($("#dvds").val() == "director name") {
             var search = $("#search").val();
             $.ajax({
                 type: "GET",
                 url: "http://localhost:8080/dvds/director/" + search,
                 contentType: "application/json;charset=UTF-8",
                 success: function (dvdArray) {
-                    $.each(dvdArray, function(index, dvd) {
+                    $.each(dvdArray, function (index, dvd) {
                         $("#dvd-results-table").append("<tr id="
                             + dvd.dvdId
-                            + "><td><button type='button' onclick=displayDVD("+dvd.dvdId+")>"
+                            + "><td><button type='button' onclick=displayDVD(" + dvd.dvdId + ")>"
                             + dvd.title
                             + "</button></td><td>"
                             + dvd.realeaseYear
@@ -99,26 +97,26 @@ $("#search-button").click( function() {
                             + dvd.director
                             + "</td><td>"
                             + dvd.rating
-                            + "</td><td><button type='button' onclick=displayEdit("+dvd.dvdId+")>Edit</button><button type='button' onclick=deleteDVD("+dvd.dvdId+")>Delete</button></td></tr>").text();
+                            + "</td><td><button type='button' onclick=displayEdit(" + dvd.dvdId + ")>Edit</button><button type='button' onclick=deleteDVD(" + dvd.dvdId + ")>Delete</button></td></tr>").text();
                     })
                 },
-                error: function(xhr, status, error) {
+                error: function (xhr, status, error) {
                     alert(xhr.responseText);
                 }
             });
         }
 
-        if($("#dvds").val() == "rating") {
+        if ($("#dvds").val() == "rating") {
             var search = $("#search").val();
             $.ajax({
                 type: "GET",
                 url: "http://localhost:8080/dvds/rating/" + search,
                 contentType: "application/json;charset=UTF-8",
                 success: function (dvdArray) {
-                    $.each(dvdArray, function(index, dvd) {
+                    $.each(dvdArray, function (index, dvd) {
                         $("#dvd-results-table").append("<tr id="
                             + dvd.dvdId
-                            + "><td><button type='button' onclick=displayDVD("+dvd.dvdId+")>"
+                            + "><td><button type='button' onclick=displayDVD(" + dvd.dvdId + ")>"
                             + dvd.title
                             + "</button></td><td>"
                             + dvd.realeaseYear
@@ -126,10 +124,10 @@ $("#search-button").click( function() {
                             + dvd.director
                             + "</td><td>"
                             + dvd.rating
-                            + "</td><td><button type='button' onclick=displayEdit("+dvd.dvdId+")>Edit</button><button type='button' onclick=deleteDVD("+dvd.dvdId+")>Delete</button></td></tr>").text();
+                            + "</td><td><button type='button' onclick=displayEdit(" + dvd.dvdId + ")>Edit</button><button type='button' onclick=deleteDVD(" + dvd.dvdId + ")>Delete</button></td></tr>").text();
                     })
                 },
-                error: function(xhr, status, error) {
+                error: function (xhr, status, error) {
                     alert(xhr.responseText);
                 }
             });
@@ -143,10 +141,10 @@ $(document).ready(function () {
         url: "http://localhost:8080/dvds",
         contentType: "application/json;charset=UTF-8",
         success: function (dvdArray) {
-            $.each(dvdArray, function(index, dvd) {
+            $.each(dvdArray, function (index, dvd) {
                 $("#dvd-results-table").append("<tr id="
                     + dvd.dvdId
-                    + "><td><button type='button' onclick=displayDVD("+dvd.dvdId+")>"
+                    + "><td><button type='button' onclick=displayDVD(" + dvd.dvdId + ")>"
                     + dvd.title
                     + "</button></td><td>"
                     + dvd.realeaseYear
@@ -154,12 +152,12 @@ $(document).ready(function () {
                     + dvd.director
                     + "</td><td>"
                     + dvd.rating
-                    + "</td><td><button type='button' onclick=displayEdit("+dvd.dvdId+")>Edit</button><button type='button' onclick=deleteDVD("+dvd.dvdId+")>Delete</button></td></tr>").text();
+                    + "</td><td><button type='button' onclick=displayEdit(" + dvd.dvdId + ")>Edit</button><button type='button' onclick=deleteDVD(" + dvd.dvdId + ")>Delete</button></td></tr>").text();
             })
         },
-        error: function(xhr, status, error) {
+        error: function (xhr, status, error) {
             alert(xhr.responseText);
-          }
+        }
     });
 });
 
@@ -181,9 +179,9 @@ function displayDVD(id) {
             $("#display-rating").append(dvd.rating).text();
             $("#display-notes").append(dvd.notes).text();
         },
-        error: function(xhr, status, error) {
+        error: function (xhr, status, error) {
             alert(xhr.responseText);
-          }
+        }
     });
 }
 
@@ -206,20 +204,68 @@ function resetDisplay() {
     $("#display-director").empty();
     $("#display-rating").empty();
     $("#display-notes").empty();
+    $("#create-release-year").empty();
+    $("#create-director").empty();
+    $("#create-rating").val() = "choose rating";
+    $("#create-notes").empty();
+    $("#edit-release-year").empty();
+    $("#edit-director").empty();
+    $("#edit-rating").val() = "choose rating";
+    $("#edit-notes").empty();
 }
 
 function deleteDVD(id) {
-    $("#confirm-delete").css("visibility", "visible");
-    $("#deleteConfirmed").click(function(id){
+    // $("#confirm-delete").css("visibility", "visible");
+    var isDelete = confirm("Are you sure you want to delete this dvd?");
+    if (isDelete) {
         $.ajax({
             type: "DELETE",
             url: "http://localhost:8080/dvd/" + id,
             contentType: "application/json;charset=UTF-8",
             success: function () {
             },
-            error: function(xhr, status, error) {
+            error: function (xhr, status, error) {
                 alert(xhr.responseText);
             }
         });
-    });
+    }
+    // $("#deleteConfirmed").click(function(id){
+    //     $.ajax({
+    //         type: "DELETE",
+    //         url: "http://localhost:8080/dvd/" + id,
+    //         contentType: "application/json;charset=UTF-8",
+    //         success: function () {
+    //         },
+    //         error: function(xhr, status, error) {
+    //             alert(xhr.responseText);
+    //         }
+    //     });
+    // });
 }
+
+$("#create-save").click(function() {
+    $.ajax({
+        type: "POST",
+        url: "http://localhost:8080/dvd",
+        data: JSON.stringify({
+            title: $("#create-title").val(),
+            releaseYear: $("#create-release-year").val(),
+            director: $("#create-director").val(),
+            rating: $("#create-rating").val(),
+            notes: $("#create-notes").val()
+        }),
+        headers: {
+            "Content-Type": "application/json",
+            "Accept": "application/json"
+        },
+        success: function () {
+        },
+        error: function (xhr, status, error) {
+            alert(xhr.responseText);
+        }
+    });
+    $("#create-release-year").empty();
+    $("#create-director").empty();
+    $("#create-rating").val() = "choose rating";
+    $("#create-notes").empty();
+});
